Simplify more the component for a generic use case
